+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Changed
+
+- Break file loading into two steps, so that the mapping step occurs after all of the
+  file data is loaded. This improves performance when there are thousands of files.
+
+
 ## [v2.1.6]
 
 ### Added